+++ src/components/AppLayout/HomeView/ConvoView/MessagesList/MessagesList.tsx
@@ -1,9 +1,9 @@
-import { Message } from '../../../../../types/appTypes';
-import Post from './Post/Post';
+import { MessageData } from '../../../../../types/appTypes';
 import styles from './MessagesList.module.css';
+import Message from './Message/Message';
 
 type MessagesListProps = {
-  data: Message[];
+  data: MessageData[];
 };
 
 function MessagesList({ data }: MessagesListProps) {
@@ -11,7 +11,7 @@ function MessagesList({ data }: MessagesListProps) {
     <div className={styles['messages-list']}>
       <div className={styles['reverse-scroll']}>
         {data.map((el) => (
-          <Post data={el} key={el.id} />
+          <Message data={el} key={el.id} />
         ))}
       </div>
     </div>
